Add files via upload
The Bislerium Café is a C# is pronounced as “see sharp” desktop application that 
helps to recognize the need of customers experiences and efficient operations with the 
implementation of Point of Sales (POS) system. Here in this application, it allows the 
admin to manage users and edit the prices of coffee and add-ins. It requires password to 
get -in. It can also generate reports and configure membership. Similarly, the staff also 
need to get log in with protected password and username. The staff can take orders, 
record the transactions, generate repots. Additionally, the program stores and retrieves 
each item's addition and removal information as well as the service center's inventory in 
a JSON file.